Site regression model

The site regression model also named genotype plus genotype-by-environment (GGE) model is powerful a tools for effective analysis and interpretation of multienvironment data structure in breeding programs.This function is a modification of GGEModel of GGEBiplots package, where the dataset contains genotype by environment means with the genotypes in rows and environments in columns. By contrast, in this function a less restrictive format of the data set is allowed, such as repetitions or other variables that will not be used in the analysis can be included.

GGEmodel(
  Data,
  genotype = "gen",
  environment = "env",
  response = "yield",
  rep = NULL,
  centering = "tester",
  scaling = "none",
  SVP = "column"
)

Arguments

Data
a data frame

genotype
name of the column that contains the genotypes

environment
name of the column that contains the environments

response
name of the column that contains the response

rep
name of the column that contains the replications.If this argument is NULL, there is no replications in the data.

centering
centering method. Either "tester" for tester centered (G+GE), "global" for global centered (E+G+GE), "double" for double centred (GE) or "none" for no centering. If a centering method is not used, the GGEPlot function can not be used.

scaling
scaling method. Either "sd" for standard deviation or "none" for no scaling.

SVP
method for singular value partitioning. Either "row","column", "dual" or "symmetrical".

Value

A list of class GGE_Model containing:

coordgenotype

plotting coordinates for genotypes from all components

coordenviroment

plotting coordinates for environments from all components

eigenvalues

vector of eigenvalues from each component

vartotal

overall variance

varexpl

percentage of variance explained by each component

labelgen

genotype names

labelenv

environment names

axes

axis labels

Data

scaled and centered input data

centering

name of centering method

scaling

name of scaling method

SVP

name of SVP method

References

Yan W, Kang M (2003). GGE Biplot Analysis: A Graphical Tool for Breeders, Geneticists, and Agronomists. CRC Press.

Yan W, Kang M (2002). Singular-Value Partitioning in Biplot Analysis of Multienvironment Trial Data. Agronomy Journal, 94, 990-996. http://dx.doi.org/10.2134/agronj2002.0990

Examples

 library(geneticae)
 # Data without replication
 GGE1 <- GGEmodel(yan.winterwheat, genotype="gen",environment="env", response="yield",
 centering = "tester")

 # Data with replication
 data(plrv)
 GGE2 <- GGEmodel(plrv, genotype="Genotype",environment="Locality",
 response="Yield", rep="Rep", centering = "tester")