Site Regression model

The Site Regression model (also called genotype + genotype-by-environment (GGE) model) is a powerful tool for effective analysis and interpretation of data from multi-environment trials in breeding programs. This function is a wrapper for GGEModel from the GGEBiplots package, with the following enhacements:

It can be used for data from trials with repetitions (there is no need to calculate means beforehand)

Other variables not used in the analysis can be present in the dataset.

Variables can be in any order in the dataset.

GGEmodel(
  Data,
  genotype = "gen",
  environment = "env",
  response = "yield",
  rep = NULL,
  centering = "tester",
  scaling = "none",
  SVP = "symmetrical"
)

Arguments

Data

dataframe with genotypes, environments, repetitions (if any) and the phenotypic trait of interest. Variables can be in any order, even additional variables that will not be used in the model may be present in the data.

genotype

column name for genotypes.

environment

column name for environments.

response

column name for the phenotypic trait.

rep

column name for replications. If this argument is NULL, there are no replications in the data. Defaults to NULL.

centering

centering method. Either `"tester"` for tester centered (G+GE), `"global"` for global centered (E+G+GE), `"double"` for double centered (GE) or `"none"` for no centering. Defaults to `"tester"`.

scaling

scaling method. Either `"sd"` for standard deviation or `"none"` for no scaling. Defaults to `"none"`.

SVP

method for singular value partitioning. Either `"row"`, `"column"`, `"dual"` or `"symmetrical"`. Defaults to `"symmetrical"`.

Value

A list of class GGE_Model containing:

coordgenotype

plotting coordinates for each genotype in every component.

coordenviroment

plotting coordinates for each environment in every component.

eigenvalues

vector of eigenvalues for each component.

vartotal

overall variance.

varexpl

percentage of variance explained by each component.

labelgen

genotype names.

labelenv

environment names.

axes

axis labels.

Data

scaled and centered input data.

centering

name of centering method.

scaling

name of scaling method.

SVP

name of SVP method.

References

Sam Dumble (2017). GGEBiplots: GGE Biplots with 'ggplot2'. R package version 0.1.1. https://CRAN.R-project.org/package=GGEBiplots

Yan W, Kang M (2003). GGE Biplot Analysis: A Graphical Tool for Breeders, Geneticists, and Agronomists. CRC Press.

Yan W, Kang M (2002). Singular-Value Partitioning in Biplot Analysis of Multienvironment Trial Data. Agronomy Journal, 94, 990-996. doi: 10.2134/agronj2002.0990

Examples

 library(geneticae)
 # Data without replication
 library(agridat)
 data(yan.winterwheat)
 GGE1 <- GGEmodel(yan.winterwheat, genotype = "gen", environment = "env",
                  response = "yield", centering = "tester")

 # Data with replication
 data(plrv)
 GGE2 <- GGEmodel(plrv, genotype = "Genotype", environment = "Locality",
                  response = "Yield", rep = "Rep", centering = "tester")