Site regression model

Site regression model also named genotype plus genotype-by-environment (GGE) model is a powerful tool for effective analysis and interpretation of multienvironment data structure in breeding programs.This function is a modification of GGEModel of GGEBiplots package, where the input dataset contains genotype by environment means with genotypes in rows and environments in columns. By contrast, this function allows a less restrictive format of dataset. Repetitions and other variables that will not be used in the analysis may be present in the dataset.

GGEmodel(
  Data,
  genotype = "gen",
  environment = "env",
  response = "yield",
  rep = NULL,
  centering = "tester",
  scaling = "none",
  SVP = "column"
)

Arguments

Data
dataframe with genotypes, environments, repetitions (if any) and phenotypic trait of interest. Variables order is indistinct, even additional variables that will not be used in the model may be present in the data.

genotype
column name containing genotypes.

environment
column name containing environments.

response
column name containing phenotypic trait.

rep
column name containing replications, if this argument is NULL, there is no replication available on the data. Default to NULL.

centering
centering method. Either "tester" for tester centered (G+GE), "global" for global centered (E+G+GE), "double" for double centred (GE) or "none" for no centering. Default to "tester".

scaling
scaling method. Either "sd" for standard deviation or "none" for no scaling. Default to "none".

SVP
method for singular value partitioning. Either "row","column", "dual" or "symmetrical". Default to "column".

Value

A list of class GGE_Model containing:

coordgenotype

plotting coordinates for genotypes from all components

coordenviroment

plotting coordinates for environments from all components

eigenvalues

vector of eigenvalues from each component

vartotal

overall variance

varexpl

percentage of variance explained by each component

labelgen

genotype names

labelenv

environment names

axes

axis labels

Data

scaled and centered input data

centering

name of centering method

scaling

name of scaling method

SVP

name of SVP method

References

Sam Dumble (2017). GGEBiplots: GGE Biplots with 'ggplot2'. R package version 0.1.1. https://CRAN.R-project.org/package=GGEBiplots

Yan W, Kang M (2003). GGE Biplot Analysis: A Graphical Tool for Breeders, Geneticists, and Agronomists. CRC Press.

Yan W, Kang M (2002). Singular-Value Partitioning in Biplot Analysis of Multienvironment Trial Data. Agronomy Journal, 94, 990-996. http://dx.doi.org/10.2134/agronj2002.0990

Examples

 library(geneticae)
 # Data without replication
 data(yan.winterwheat)
 GGE1 <- GGEmodel(yan.winterwheat, genotype="gen",environment="env", response="yield",
 centering = "tester")

 # Data with replication
 data(plrv)
 GGE2 <- GGEmodel(plrv, genotype="Genotype",environment="Locality",
 response="Yield", rep="Rep", centering = "tester")