Produces genotype plus genotype-by-environment model from a 2-way table of means

Produces genotype plus genotype-by-environment model from a 2-way table of means

imputation(
  Data,
  genotype = "gen",
  environment = "env",
  response = "yield",
  rep = NULL,
  type = "EM-AMMI",
  PC.nb = 1,
  initial.values = NA,
  precision = 0.01,
  max.iter = 1000,
  change.factor = 1,
  simplified.model = FALSE,
  k = min(nrow(Data), ncol(Data)),
  tol = max(nrow(Data), ncol(Data)) * 1e-10,
  maxiter = 100,
  Winf = 0.8,
  Wsup = 1,
  ncp = 2,
  scale = TRUE,
  method = c("Regularized", "EM"),
  row.w = NULL,
  coeff.ridge = 1,
  threshold = 1e-06,
  seed = NULL,
  nb.init = 1
)

Arguments

Data
a data frame

genotype
name of the column that contains the genotypes

environment
name of the column that contains the environments

response
name of the column that contains the response

rep
name of the column that contains the replications.If this argument is NULL, there is no replications in the data.

type
imputation method. Either "EM-AMMI", "EM-SVD", "Gabriel","WGabriel","EM-PCA".

PC.nb
the number of principal components in the AMMI model that will be used; the default value is 1. For PC.nb=0 only main effects are used to estimate cells in the data table (the interaction is ignored). The number of principal components must not be greater than min(number of rows in the X table, number of columns in the X table)–2.

initial.values
(optional) the initial values of the missing cells. It can be a single value, which then will be used for all empty cells, or a vector of length equal to the number of missing cells (starting from the missing values in the first column). If omitted, the initial values will be obtained by the main effects from the corresponding model, that is, by the grand mean of the observed data increased (or decreased) by row and column main effects.

precision
(optional) the algorithm converges if the maximal change in the values of the missing cells in two subsequent steps is not greater than this value (the default is 0.01);

max.iter
(optional) a maximum permissible number of iterations (that is, number of repeats of the algorithm’s steps 2 through 5); the default value is 1000;

change.factor
(optional) introduced by analogy to step size in gradient descent method, this parameter that can shorten the time of executing the algorithm by decreasing the number of iterations. The change.factor=1 (default) defines that the previous approximation is changed with the new values of missing cells (standard EM-AMMI algorithm). However, when change.factor<1, then the new approximations are computed and the values of missing cells are changed in the direction of this new approximation but the change is smaller. It could be useful if the changes are cyclic and thus convergence could not be reached. Usually, this argument should not affect the final outcome (that is, the imputed values) as compared to the default value of change.factor=1.

simplified.model
the AMMI model contains the general mean, effects of rows, columns and interaction terms. So the EM-AMMI algorithm in step 2 calculates the current effects of rows and columns; these effects change from iteration to iteration because the empty (at the outset) cells in each iteration are filled with different values. In step 3 EM-AMMI uses those effects to re-estimate cells marked as missed (as default, simplified.model=FALSE). It is, however, possible that this procedure will not converge. Thus the user is offered a simplified EM-AMMI procedure that calculates the general mean and effects of rows and columns only in the first iteration and in next iterations uses these values (simplified.model=TRUE). In this simplified procedure the initial values affect the outcome (whilst EM-AMMI results usually do not depend on initial values). For the simplified procedure the number of iterations to convergence is usually smaller and, furthermore, convergence will be reached even in some cases where the regular procedure fails. If the regular procedure does not converge for the standard initial values (see the description of the argument initial.values), the simplified model can be used to determine a better set of initial values.

k
the rank of the SVD approximation.

tol
the convergence tolerance for the EM algorithm.

maxiter
the maximum number of EM steps to take.

Winf
peso inferior

Wsup
peso superior

ncp
integer corresponding to the number of components used to to predict the missing entries

scale
boolean. By default TRUE leading to a same weight for each variable

method
"Regularized" by default or "EM"

row.w
row weights (by default, a vector of 1 for uniform row weights)

coeff.ridge
1 by default to perform the regularized imputePCA algorithm; useful only if method="Regularized". Other regularization terms can be implemented by setting the value to less than 1 in order to regularized less (to get closer to the results of the EM method) or more than 1 to regularized more (to get closer to the results of the mean imputation)

threshold
the threshold for assessing convergence

seed
integer, by default seed = NULL implies that missing values are initially imputed by the mean of each variable. Other values leads to a random initialization

nb.init
integer corresponding to the number of random initializations; the first initialization is the initialization with the mean imputation

Value

matrix with imputed data

Examples

library(geneticae)
# Data without replication
data(yan.winterwheat)
dat <- yan.winterwheat
# generates missing data
dat[1,3]<-NA
dat[3,3]<-NA
dat[2,3]<-NA
imputation(dat, genotype="gen",environment="env", response="yield", type="EM-SVD")

#>           [,1]  [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9]
#>  [1,] 4.332467 4.150 2.849 3.084 5.940 4.450 4.351 4.039 2.672
#>  [2,] 4.332467 4.771 2.912 3.506 5.699 5.152 4.956 4.386 2.938
#>  [3,] 4.332467 4.578 3.098 3.460 6.070 5.025 4.730 3.900 2.621
#>  [4,] 4.732000 4.745 3.375 3.904 6.224 5.340 4.226 4.893 3.451
#>  [5,] 4.390000 4.603 3.511 3.848 5.773 5.421 5.147 4.098 2.832
#>  [6,] 5.178000 4.475 2.990 3.774 6.583 5.045 3.985 4.271 2.776
#>  [7,] 3.375000 4.175 2.741 3.157 5.342 4.267 4.162 4.063 2.032
#>  [8,] 4.852000 4.664 4.425 3.952 5.536 5.832 4.168 5.060 3.574
#>  [9,] 5.038000 4.741 3.508 3.437 5.960 4.859 4.977 4.514 2.859
#> [10,] 5.195000 4.662 3.596 3.759 5.937 5.345 3.895 4.450 3.300
#> [11,] 4.293000 4.530 2.760 3.422 6.142 5.250 4.856 4.137 3.149
#> [12,] 3.151000 3.040 2.388 2.350 4.229 4.257 3.384 4.071 2.103
#> [13,] 4.104000 3.878 2.302 3.718 4.555 5.149 2.596 4.956 2.886
#> [14,] 4.375000 4.701 3.655 3.592 6.189 5.141 3.933 4.208 2.925
#> [15,] 4.940000 4.698 2.950 3.898 6.063 5.326 4.302 4.299 3.031
#> [16,] 3.786000 4.969 3.379 3.353 4.774 5.304 4.322 4.858 3.382
#> [17,] 4.238000 4.654 3.607 3.914 6.641 4.830 5.014 4.363 3.111
#> [18,] 3.340000 3.854 2.419 2.783 4.629 5.090 3.281 3.918 2.561

imputation(dat, genotype="gen",environment="env", response="yield", type="EM-AMMI")

#>         BH93  EA93  HW93  ID93  KE93  NN93  OA93  RN93  WP93
#> Ann 4.136249 4.150 2.849 3.084 5.940 4.450 4.351 4.039 2.672
#> Ari 4.474249 4.771 2.912 3.506 5.699 5.152 4.956 4.386 2.938
#> Aug 4.386299 4.578 3.098 3.460 6.070 5.025 4.730 3.900 2.621
#> Cas 4.732000 4.745 3.375 3.904 6.224 5.340 4.226 4.893 3.451
#> Del 4.390000 4.603 3.511 3.848 5.773 5.421 5.147 4.098 2.832
#> Dia 5.178000 4.475 2.990 3.774 6.583 5.045 3.985 4.271 2.776
#> Ena 3.375000 4.175 2.741 3.157 5.342 4.267 4.162 4.063 2.032
#> Fun 4.852000 4.664 4.425 3.952 5.536 5.832 4.168 5.060 3.574
#> Ham 5.038000 4.741 3.508 3.437 5.960 4.859 4.977 4.514 2.859
#> Har 5.195000 4.662 3.596 3.759 5.937 5.345 3.895 4.450 3.300
#> Kar 4.293000 4.530 2.760 3.422 6.142 5.250 4.856 4.137 3.149
#> Kat 3.151000 3.040 2.388 2.350 4.229 4.257 3.384 4.071 2.103
#> Luc 4.104000 3.878 2.302 3.718 4.555 5.149 2.596 4.956 2.886
#> Reb 4.375000 4.701 3.655 3.592 6.189 5.141 3.933 4.208 2.925
#> Ron 4.940000 4.698 2.950 3.898 6.063 5.326 4.302 4.299 3.031
#> Rub 3.786000 4.969 3.379 3.353 4.774 5.304 4.322 4.858 3.382
#> Zav 4.238000 4.654 3.607 3.914 6.641 4.830 5.014 4.363 3.111
#> m12 3.340000 3.854 2.419 2.783 4.629 5.090 3.281 3.918 2.561

imputation(dat, genotype="gen",environment="env", response="yield", type="Gabriel")

#>         BH93  EA93  HW93  ID93  KE93  NN93  OA93  RN93  WP93
#> Ann 3.982510 4.150 2.849 3.084 5.940 4.450 4.351 4.039 2.672
#> Ari 4.462387 4.771 2.912 3.506 5.699 5.152 4.956 4.386 2.938
#> Aug 4.325113 4.578 3.098 3.460 6.070 5.025 4.730 3.900 2.621
#> Cas 4.732000 4.745 3.375 3.904 6.224 5.340 4.226 4.893 3.451
#> Del 4.390000 4.603 3.511 3.848 5.773 5.421 5.147 4.098 2.832
#> Dia 5.178000 4.475 2.990 3.774 6.583 5.045 3.985 4.271 2.776
#> Ena 3.375000 4.175 2.741 3.157 5.342 4.267 4.162 4.063 2.032
#> Fun 4.852000 4.664 4.425 3.952 5.536 5.832 4.168 5.060 3.574
#> Ham 5.038000 4.741 3.508 3.437 5.960 4.859 4.977 4.514 2.859
#> Har 5.195000 4.662 3.596 3.759 5.937 5.345 3.895 4.450 3.300
#> Kar 4.293000 4.530 2.760 3.422 6.142 5.250 4.856 4.137 3.149
#> Kat 3.151000 3.040 2.388 2.350 4.229 4.257 3.384 4.071 2.103
#> Luc 4.104000 3.878 2.302 3.718 4.555 5.149 2.596 4.956 2.886
#> Reb 4.375000 4.701 3.655 3.592 6.189 5.141 3.933 4.208 2.925
#> Ron 4.940000 4.698 2.950 3.898 6.063 5.326 4.302 4.299 3.031
#> Rub 3.786000 4.969 3.379 3.353 4.774 5.304 4.322 4.858 3.382
#> Zav 4.238000 4.654 3.607 3.914 6.641 4.830 5.014 4.363 3.111
#> m12 3.340000 3.854 2.419 2.783 4.629 5.090 3.281 3.918 2.561

imputation(dat, genotype="gen",environment="env", response="yield", type="WGabriel")

#> [1] "  PesoExterno =  0.8"
#> [1] "  PesoExterno =  0.9"
#> [1] "  PesoExterno =  1"
#> [1] "    PesoInterno =  0.8"
#> [1] "    PesoInterno =  0.81"
#> [1] "    PesoInterno =  0.82"
#> [1] "    PesoInterno =  0.83"
#> [1] "    PesoInterno =  0.84"
#> [1] "    PesoInterno =  0.85"
#> [1] "    PesoInterno =  0.86"
#> [1] "    PesoInterno =  0.87"
#> [1] "    PesoInterno =  0.88"
#> [1] "    PesoInterno =  0.89"
#> [1] "    PesoInterno =  0.9"
#> [1] "    PesoInterno =  0.91"
#> [1] "    PesoInterno =  0.92"
#> [1] "    PesoInterno =  0.93"
#> [1] "    PesoInterno =  0.94"
#> [1] "    PesoInterno =  0.95"
#> [1] "    PesoInterno =  0.96"
#> [1] "    PesoInterno =  0.97"
#> [1] "    PesoInterno =  0.98"
#> [1] "    PesoInterno =  0.99"
#> [1] "    PesoInterno =  1"

#>         BH93  EA93  HW93  ID93  KE93  NN93  OA93  RN93  WP93
#> Ann 4.004664 4.150 2.849 3.084 5.940 4.450 4.351 4.039 2.672
#> Ari 4.455727 4.771 2.912 3.506 5.699 5.152 4.956 4.386 2.938
#> Aug 4.328095 4.578 3.098 3.460 6.070 5.025 4.730 3.900 2.621
#> Cas 4.732000 4.745 3.375 3.904 6.224 5.340 4.226 4.893 3.451
#> Del 4.390000 4.603 3.511 3.848 5.773 5.421 5.147 4.098 2.832
#> Dia 5.178000 4.475 2.990 3.774 6.583 5.045 3.985 4.271 2.776
#> Ena 3.375000 4.175 2.741 3.157 5.342 4.267 4.162 4.063 2.032
#> Fun 4.852000 4.664 4.425 3.952 5.536 5.832 4.168 5.060 3.574
#> Ham 5.038000 4.741 3.508 3.437 5.960 4.859 4.977 4.514 2.859
#> Har 5.195000 4.662 3.596 3.759 5.937 5.345 3.895 4.450 3.300
#> Kar 4.293000 4.530 2.760 3.422 6.142 5.250 4.856 4.137 3.149
#> Kat 3.151000 3.040 2.388 2.350 4.229 4.257 3.384 4.071 2.103
#> Luc 4.104000 3.878 2.302 3.718 4.555 5.149 2.596 4.956 2.886
#> Reb 4.375000 4.701 3.655 3.592 6.189 5.141 3.933 4.208 2.925
#> Ron 4.940000 4.698 2.950 3.898 6.063 5.326 4.302 4.299 3.031
#> Rub 3.786000 4.969 3.379 3.353 4.774 5.304 4.322 4.858 3.382
#> Zav 4.238000 4.654 3.607 3.914 6.641 4.830 5.014 4.363 3.111
#> m12 3.340000 3.854 2.419 2.783 4.629 5.090 3.281 3.918 2.561

imputation(dat, genotype="gen",environment="env", response="yield", type="EM-PCA")

#>         BH93  EA93  HW93  ID93  KE93  NN93  OA93  RN93  WP93
#> Ann 3.980317 4.150 2.849 3.084 5.940 4.450 4.351 4.039 2.672
#> Ari 4.463093 4.771 2.912 3.506 5.699 5.152 4.956 4.386 2.938
#> Aug 4.327731 4.578 3.098 3.460 6.070 5.025 4.730 3.900 2.621
#> Cas 4.732000 4.745 3.375 3.904 6.224 5.340 4.226 4.893 3.451
#> Del 4.390000 4.603 3.511 3.848 5.773 5.421 5.147 4.098 2.832
#> Dia 5.178000 4.475 2.990 3.774 6.583 5.045 3.985 4.271 2.776
#> Ena 3.375000 4.175 2.741 3.157 5.342 4.267 4.162 4.063 2.032
#> Fun 4.852000 4.664 4.425 3.952 5.536 5.832 4.168 5.060 3.574
#> Ham 5.038000 4.741 3.508 3.437 5.960 4.859 4.977 4.514 2.859
#> Har 5.195000 4.662 3.596 3.759 5.937 5.345 3.895 4.450 3.300
#> Kar 4.293000 4.530 2.760 3.422 6.142 5.250 4.856 4.137 3.149
#> Kat 3.151000 3.040 2.388 2.350 4.229 4.257 3.384 4.071 2.103
#> Luc 4.104000 3.878 2.302 3.718 4.555 5.149 2.596 4.956 2.886
#> Reb 4.375000 4.701 3.655 3.592 6.189 5.141 3.933 4.208 2.925
#> Ron 4.940000 4.698 2.950 3.898 6.063 5.326 4.302 4.299 3.031
#> Rub 3.786000 4.969 3.379 3.353 4.774 5.304 4.322 4.858 3.382
#> Zav 4.238000 4.654 3.607 3.914 6.641 4.830 5.014 4.363 3.111
#> m12 3.340000 3.854 2.419 2.783 4.629 5.090 3.281 3.918 2.561

# Data with replication
 data(plrv)
 dat2<-plrv
 dat2[1,3]<-NA
dat2[3,3]<-NA
dat2[2,3]<-NA
imputation(dat2, genotype="gen",environment="env", response="yield", rep= "rep", type="EM-SVD")

#> Error: Column `gen` is unknown

imputation(dat2, genotype="gen",environment="env", response="yield", rep= "rep", type="EM-AMMI")

#> Error: Column `gen` is unknown

imputation(dat2, genotype="gen",environment="env", response="yield", rep= "rep", type="Gabriel")

#> Error: Column `gen` is unknown

imputation(dat2, genotype="gen",environment="env", response="yield", rep= "rep", type="WGabriel")

#> Error: Column `gen` is unknown

imputation(dat2, genotype="gen",environment="env", response="yield", rep= "rep", type="EM-PCA")

#> Error: Column `gen` is unknown